Fix: request id

+++ src/parser/models/session.py
@@ -3,5 +3,5 @@ from .request import Request_Model
 
 
 class Session_Model(Base_model):
-    session_id: float | None = None
+    session_id: int | None = None
     requests: list[Request_Model] = []
